Move tests to components folders
