+++ src/components/ButtonIcon/__tests__/ButtonIcon.test.js
@@ -3,7 +3,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 
-import ButtonIcon from '../src/components/ButtonIcon';
+import ButtonIcon from '../ButtonIcon';
 
 it('renders a ButtonIcon', () => {
   expect(
