+++ src/components/NotificationIcon/__tests__/NotificationIcon.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import NotificationIcon from '../src/components/NotificationIcon';
+import NotificationIcon from '../NotificationIcon';
 
 it('renders a NotificationIcon', () => {
   expect(renderer.create(<NotificationIcon count={5} showShadow />)).toMatchSnapshot();
